requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
PyMySQL==1.1.1
itsdangerous==2.2.0
MarkupSafe==3.0.2
Werkzeug==3.1.3
gunicorn==23.0.0
python-dotenv==1.0.0
cryptography==43.0.3
